Separate Command and Completion reducer
