+++ src/console/components/AppContext.ts
@@ -1,6 +1,6 @@
 import React from "react";
-import { State, defaultState } from "../reducers";
-import { ConsoleAction } from "../actions";
+import { State, defaultState } from "../reducers/console";
+import { ConsoleAction } from "../actions/console";
 
 const AppContext = React.createContext<{
   state: State;
