+++ test/console/reducers/console.test.ts
@@ -1,8 +1,14 @@
-import * as actions from "../../../src/console/actions";
-import reducer, { State } from "../../../src/console/reducers";
+import reducer from "../../../src/console/reducers/console";
 import { expect } from "chai";
-import CompletionType from "../../../src/shared/CompletionType";
-import { ConsoleAction } from "../../../src/console/actions";
+import {
+  CONSOLE_HIDE,
+  CONSOLE_HIDE_COMMAND,
+  CONSOLE_SET_CONSOLE_TEXT,
+  CONSOLE_SHOW_COMMAND,
+  CONSOLE_SHOW_ERROR,
+  CONSOLE_SHOW_INFO,
+  ConsoleAction,
+} from "../../../src/console/actions/console";
 
 describe("console reducer", () => {
   it("return the initial state", () => {
@@ -10,21 +16,18 @@ describe("console reducer", () => {
     expect(state).to.have.property("mode", "");
     expect(state).to.have.property("messageText", "");
     expect(state).to.have.property("consoleText", "");
-    expect(state).to.have.deep.property("completions", []);
-    expect(state).to.have.property("select", -1);
   });
 
   it("return next state for CONSOLE_HIDE", () => {
     const initialState = reducer(undefined, {} as any);
-    const action: actions.ConsoleAction = { type: actions.CONSOLE_HIDE };
+    const action: ConsoleAction = { type: CONSOLE_HIDE };
     const state = reducer({ ...initialState, mode: "error" }, action);
     expect(state).to.have.property("mode", "");
   });
 
   it("return next state for CONSOLE_SHOW_COMMAND", () => {
-    const action: actions.ConsoleAction = {
-      type: actions.CONSOLE_SHOW_COMMAND,
-      completionTypes: [CompletionType.SearchEngines, CompletionType.History],
+    const action: ConsoleAction = {
+      type: CONSOLE_SHOW_COMMAND,
       text: "open ",
     };
     const state = reducer(undefined, action);
@@ -33,8 +36,8 @@ describe("console reducer", () => {
   });
 
   it("return next state for CONSOLE_SHOW_INFO", () => {
-    const action: actions.ConsoleAction = {
-      type: actions.CONSOLE_SHOW_INFO,
+    const action: ConsoleAction = {
+      type: CONSOLE_SHOW_INFO,
       text: "an info",
     };
     const state = reducer(undefined, action);
@@ -43,8 +46,8 @@ describe("console reducer", () => {
   });
 
   it("return next state for CONSOLE_SHOW_ERROR", () => {
-    const action: actions.ConsoleAction = {
-      type: actions.CONSOLE_SHOW_ERROR,
+    const action: ConsoleAction = {
+      type: CONSOLE_SHOW_ERROR,
       text: "an error",
     };
     const state = reducer(undefined, action);
@@ -54,8 +57,8 @@ describe("console reducer", () => {
 
   it("return next state for CONSOLE_HIDE_COMMAND", () => {
     const initialState = reducer(undefined, {} as any);
-    const action: actions.ConsoleAction = {
-      type: actions.CONSOLE_HIDE_COMMAND,
+    const action: ConsoleAction = {
+      type: CONSOLE_HIDE_COMMAND,
     };
     let state = reducer({ ...initialState, mode: "command" }, action);
     expect(state).to.have.property("mode", "");
@@ -65,100 +68,12 @@ describe("console reducer", () => {
   });
 
   it("return next state for CONSOLE_SET_CONSOLE_TEXT", () => {
-    const action: actions.ConsoleAction = {
-      type: actions.CONSOLE_SET_CONSOLE_TEXT,
+    const action: ConsoleAction = {
+      type: CONSOLE_SET_CONSOLE_TEXT,
       consoleText: "hello world",
     };
     const state = reducer(undefined, action);
 
     expect(state).to.have.property("consoleText", "hello world");
-  });
-
-  it("return next state for CONSOLE_SET_COMPLETIONS", () => {
-    const initialState = reducer(undefined, {} as any);
-    let state: State = {
-      ...initialState,
-      select: 0,
-      completions: [],
-    };
-    const action: actions.ConsoleAction = {
-      type: actions.CONSOLE_SET_COMPLETIONS,
-      completions: [
-        {
-          name: "Apple",
-          items: [{}, {}, {}],
-        },
-        {
-          name: "Banana",
-          items: [{}, {}, {}],
-        },
-      ],
-      completionSource: "",
-    };
-    state = reducer(state, action);
-    expect(state).to.have.property("completions", action.completions);
-    expect(state).to.have.property("select", -1);
-  });
-
-  it("return next state for CONSOLE_COMPLETION_NEXT", () => {
-    const initialState = reducer(undefined, {} as any);
-    const action: ConsoleAction = { type: actions.CONSOLE_COMPLETION_NEXT };
-    let state = {
-      ...initialState,
-      select: -1,
-      completions: [
-        {
-          name: "Apple",
-          items: [{}, {}],
-        },
-        {
-          name: "Banana",
-          items: [{}],
-        },
-      ],
-    };
-
-    state = reducer(state, action);
-    expect(state).to.have.property("select", 0);
-
-    state = reducer(state, action);
-    expect(state).to.have.property("select", 1);
-
-    state = reducer(state, action);
-    expect(state).to.have.property("select", 2);
-
-    state = reducer(state, action);
-    expect(state).to.have.property("select", -1);
-  });
-
-  it("return next state for CONSOLE_COMPLETION_PREV", () => {
-    const initialState = reducer(undefined, {} as any);
-    const action: ConsoleAction = { type: actions.CONSOLE_COMPLETION_PREV };
-    let state = {
-      ...initialState,
-      select: -1,
-      completions: [
-        {
-          name: "Apple",
-          items: [{}, {}],
-        },
-        {
-          name: "Banana",
-          items: [{}],
-        },
-      ],
-    };
-
-    state = reducer(state, action);
-    expect(state).to.have.property("select", 2);
-
-    state = reducer(state, action);
-    expect(state).to.have.property("select", 1);
-
-    state = reducer(state, action);
-    expect(state).to.have.property("select", 0);
-
-    state = reducer(state, action);
-    expect(state).to.have.property("select", -1);
   });
 });
